Add filter and to be canceled it on FeedList
詳細ページでフィルタ追加した際の動きをいかに変更
- 自動的に記事一覧に戻る
- 記事一覧ページで追加したフィルタをSnackbarで表示
- フィルタの追加をキャンセル可能に
この仕様変更に伴い、Filterの追加・削除に関する処理はFilterServiceで行うようにした

+++ app/src/main/java/jp/gcreate/product/filteredhatebu/domain/services/FilterService.kt
@@ -1,9 +1,12 @@
 package jp.gcreate.product.filteredhatebu.domain.services
 
+import android.arch.lifecycle.LiveData
+import android.arch.lifecycle.MutableLiveData
 import jp.gcreate.product.filteredhatebu.data.AppRoomDatabase
 import jp.gcreate.product.filteredhatebu.data.entities.FeedFilter
 import jp.gcreate.product.filteredhatebu.data.entities.FilteredFeed
 import jp.gcreate.product.filteredhatebu.di.Scope.AppScope
+import jp.gcreate.product.filteredhatebu.ui.common.HandleOnceEvent
 import kotlinx.coroutines.experimental.CommonPool
 import kotlinx.coroutines.experimental.async
 import kotlinx.coroutines.experimental.launch
@@ -19,9 +22,15 @@ class FilterService @Inject constructor(
     private val feedFilterDao = appRoomDatabase.feedFilterDao()
     private val feedDataDao = appRoomDatabase.feedDataDao()
     private val filteredFeedDao = appRoomDatabase.filteredFeedDao()
+    private var deleteCommand: DeleteCommand? = null
+    private var addedFilter: String? = null
+    private val addFilterEventEmitter = MutableLiveData<HandleOnceEvent<FeedFilter>>()
+    val addFilterEvent: LiveData<HandleOnceEvent<FeedFilter>> = addFilterEventEmitter
     
     fun addFilter(filter: String) = launch(CommonPool) {
-        val ids = feedFilterDao.insertFilter(FeedFilter(0, filter, ZonedDateTime.now()))
+        addedFilter = filter
+        val insert = FeedFilter(0, filter, ZonedDateTime.now())
+        val ids = feedFilterDao.insertFilter(insert)
         val id = ids[0]
         
         if (id == 0L) {
@@ -34,11 +43,41 @@ class FilterService @Inject constructor(
                 async { filteredFeedDao.insertFilteredFeed(FilteredFeed(id, it.url)) }
             }
         }
+        addFilterEventEmitter.postValue(HandleOnceEvent(insert))
+    }
+    
+    fun undoAdd() = launch(CommonPool) {
+        addedFilter?.let {
+            feedFilterDao.deleteFilter(it)
+            addedFilter = null
+        }
     }
     
     fun deleteFilter(filter: String) {
         launch {
+            val target = feedFilterDao.getFilter(filter)
+            if (target == null) {
+                Timber.e("$filter is not exist")
+                return@launch
+            }
+            val feeds = filteredFeedDao.getFilteredFeed(target.id)
+            deleteCommand = DeleteCommand(target, feeds)
             feedFilterDao.deleteFilter(filter)
         }
     }
+    
+    fun undoDelete() {
+        deleteCommand?.let {
+            feedFilterDao.insertFilter(it.filter)
+            if (it.filteredFeeds.isNotEmpty()) {
+                filteredFeedDao.insertFilteredFeed(*it.filteredFeeds.toTypedArray())
+            }
+            deleteCommand = null
+        }
+    }
 }
+
+data class DeleteCommand(
+    val filter: FeedFilter,
+    val filteredFeeds: List<FilteredFeed>
+)
